Remove warning about unused local variable.
This removes the warning that was occuring by passing the `HTMLFormatter#binding`
to a erb template without using that variable in the method.

The binding given to the erb template is that of a basic object that has no
methods or variables apart from accessor methods. This hides the internals of
HTMLFormatter from the ERB template.

This change was tested locally against the cucumber-ruby-core project.

Please consider a 0.8.1 release as this warning has propergated down to other
projects that depend on simplecov-html.