Revert "Fix UTF8 encoding error"
This reverts commit 2608b9db2f2012c1223e88b0275e48ef10818553.

Replaced by: https://github.com/colszowka/simplecov/pull/866
As suggested by: https://github.com/colszowka/simplecov-html/issues/91#issuecomment-586101265
Should be safe to revert as apparently ruby itself doesn't work
when the encoding isn't declared properly so should be good.

For files that aren't required by ruby but tracked we added a
test case with broken encoding (in simplecov main) but it still
seems to work.  So this complexity is likely not needed anymore.