Show Branch Coverage in stdout if enabled.
Prior to this only Line Coverage is shown in the stdout. 

Branch coverage is equally important, when it is enabled, so it should be shown alongside it to ensure adequate coverage of the code base.

Now it shows in a clean format, including covered branches, total branches and percent of covered branches. Something like this: 

```
Coverage report generated for (1/10), (10/10), (2/10), (3/10), (4/10), (5/10), (6/10), (7/10), (8/10), (9/10) to /Users/me/Development/cntral/coverage.
Line Coverage: 7554 / 12927 (58.44%)
Branch Coverage: 1868 / 3704 (50.43%)
```

When branch coverage is not enabled it simply omits that line, so it appears like this: 

```
Coverage report generated for (1/10), (10/10), (2/10), (3/10), (4/10), (5/10), (6/10), (7/10), (8/10), (9/10) to /Users/me/Development/cntral/coverage.
Line Coverage: 7554 / 12927 (58.44%)
```